refactor: remove old play route and host specific route to play game workflow
