--- pages/host/[joinCode]/[round]/overview.jsx
@@ -1,1 +1,0 @@
-import * as React from 'react'
