--- pages/host/[joinCode]/overview.jsx
@@ -1,1 +1,0 @@
-import * as React from 'react'
